+++ app/models/department.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Annotated
 from pydantic import BaseModel, BeforeValidator
 from app.models import default_model_config, FieldAlias
@@ -6,7 +7,7 @@ from app.models import default_model_config, FieldAlias
 class Department(BaseModel):
     id: str = FieldAlias("DepartmentID")
     name: str = FieldAlias("Name")
-    budget: float = FieldAlias("Budget")
+    budget: Decimal = FieldAlias("Budget")
     created_at: Annotated[int, BeforeValidator(
         lambda x: int(x))] = FieldAlias("CreatedAt", default=0)
     updated_at: Annotated[int, BeforeValidator(
