+++ app/models/project.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Annotated
 from pydantic import BaseModel, BeforeValidator
 from app.models import default_model_config, FieldAlias
@@ -7,14 +8,13 @@ class Project(BaseModel):
     id: str = FieldAlias("ProjectID")
     name: str = FieldAlias("Name")
     description: str = FieldAlias("Description")
-    budget: float = FieldAlias("Budget")
+    budget: Decimal = FieldAlias("Budget")
     start_date: int = FieldAlias("StartDate")
     end_date: int = FieldAlias("EndDate")
     created_at: Annotated[int, BeforeValidator(
         lambda x: int(x))] = FieldAlias("CreatedAt", default=0)
     updated_at: Annotated[int, BeforeValidator(
         lambda x: int(x))] = FieldAlias("UpdatedAt", default=0)
-
 
     # pydantic config
     model_config = default_model_config()
